+++ music_therapy.rb
@@ -1,8 +1,10 @@
 require "roda"
 require 'rspotify'
 require 'cgi'
-require_relative 'db'
 require 'faraday'
+
+require_relative 'db'
+require_relative 'models'
 
 class MusicTherapy < Roda
 
@@ -10,7 +12,7 @@ class MusicTherapy < Roda
   plugin :public
   plugin :render
 
-  HOSTNAME = ENV["RACK_HOST"] ? ENV["RACK_HOST"] : "http://localhost:9292"
+  HOSTNAME = ( ENV["RACK_HOST"] || "http://localhost:9292" )
 
   CLIENT_ID = ENV["CLIENT_ID"]
   CLIENT_SECRET = ENV["CLIENT_SECRET"]
